+++ djangocms_blog/models.py
@@ -88,8 +88,8 @@ class BlogCategory(BlogMetaMixin, TranslatableModel):
     )
 
     translations = TranslatedFields(
-        name=models.CharField(_('name'), max_length=767),
-        slug=models.SlugField(_('slug'), max_length=767, blank=True, db_index=True),
+        name=models.CharField(_('name'), max_length=752),
+        slug=models.SlugField(_('slug'), max_length=752, blank=True, db_index=True),
         meta_description=models.TextField(
             verbose_name=_('category meta description'), blank=True, default=''
         ),
@@ -224,8 +224,8 @@ class Post(KnockerModel, BlogMetaMixin, TranslatableModel):
     )
 
     translations = TranslatedFields(
-        title=models.CharField(_('title'), max_length=767),
-        slug=models.SlugField(_('slug'), max_length=767, blank=True,
+        title=models.CharField(_('title'), max_length=752),
+        slug=models.SlugField(_('slug'), max_length=752, blank=True,
                               db_index=True, allow_unicode=True),
         subtitle=models.CharField(verbose_name=_('subtitle'), max_length=767,
                                   blank=True, default=''),