+++ djangocms_blog/__init__.py
@@ -3,6 +3,6 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 __author__ = 'Iacopo Spalletti'
 __email__ = 'i.spalletti@nephila.it'
-__version__ = '1.0.1.dev1'
+__version__ = '1.1.0'
 
 default_app_config = 'djangocms_blog.apps.BlogAppConfig'