+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 *******************
 
 * Remove cmsplugin_filer from installation docs
+* Set minimum django-parler version to 2.0
 
 *******************
 1.0.0 (2019-11-04)