+++ HISTORY.rst
@@ -10,6 +10,8 @@ History
 
 * Remove cmsplugin_filer from installation docs
 * Set minimum django-parler version to 2.0
+* Fix to respect current locale for category names
+* Reduce the maximum length of slug fields to 752 characters
 
 *******************
 1.0.0 (2019-11-04)